[update]:詳細ページのAppRouter化（前後リンクはコメントアウト）
